# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "My First API",
        "version": "0.1"
    },
    "paths": {
...
        "/api/v1/categories": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách tất cả các danh mục",
                "operationId": "getCategories",
                "responses": {
                    "200": {
                        "description": "Danh sách tất cả các danh mục"
                    }
                }
            },
            "post": {
                "tags": [
                    "Categories"
                ],
                "summary": "Tạo danh mục mới",
                "operationId": "createCategory",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Tên của danh mục",
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "description": "ID của danh mục cha (nếu có)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Danh mục mới đã được tạo"
                    }
                }
            }
        },
        "/api/v1/categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy thông tin của danh mục theo ID",
                "operationId": "getCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Thông tin của danh mục"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            },
            "put": {
                "tags": [
                    "Categories"
                ],
                "summary": "Cập nhật danh mục theo ID",
                "operationId": "updateCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "name": {
                                        "description": "Tên của danh mục",
                                        "type": "string"
                                    },
                                    "parent_id": {
                                        "description": "ID của danh mục cha (nếu có)",
                                        "type": "integer"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Danh mục đã được cập nhật"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Categories"
                ],
                "summary": "Xóa danh mục theo ID",
                "operationId": "deleteCategory",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Danh mục đã được xóa"
                    },
                    "404": {
                        "description": "Không tìm thấy danh mục"
                    }
                }
            }
        },
...
        "/api/v1/sub-categories/{id}": {
            "get": {
                "tags": [
                    "Categories"
                ],
                "summary": "Lấy danh sách các danh mục con dựa trên ID danh mục cha",
                "operationId": "getSubCategories",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID của danh mục cha",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Danh sách các danh mục con"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}